+++ src/Resources/Resources/DenyAssignments/GetAzureDenyAssignmentCommand.cs
@@ -17,7 +17,7 @@ using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using Microsoft.Azure.Commands.Resources.Models;
 using Microsoft.Azure.Commands.Resources.Models.Authorization;
 using Microsoft.WindowsAzure.Commands.Common;
-using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
+
 using System;
 using System.Collections.Generic;
 using System.Management.Automation;
@@ -27,7 +27,6 @@ namespace Microsoft.Azure.Commands.Resources
     /// <summary>
     /// Filters deny assignments
     /// </summary>
-    [GenericBreakingChange(BreakingChangeMSGraphMigration)]
     [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "DenyAssignment", DefaultParameterSetName = ParameterSet.Empty), OutputType(typeof(PSDenyAssignment))]
     public class GetAzureDenyAssignmentCommand : ResourcesBaseCmdlet
     {