+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Changed the Scope Property in New-AzRoleAssignment docs. 
 
 ## Version 4.4.1
 * Fixed a bug about the exitcode of Bicep [#16055]